ci: add ci-artifacts pipeline
A similar pipeline already exists in git-sdk-64, so this commit is to port most of that code to git-sdk-arm64.

We don't (yet) copy over the test-minimal-sdk job, because it spins up 16 parallel jobs, which is a bit too much for our self-hosted runner budget. We can add those once GitHub-hosted arm64 runners become available for OSS projects by the end of 2024.

Ref: https://github.com/git-for-windows/git-for-windows-automation/issues/91
Ref: https://github.blog/news-insights/product-news/arm64-on-github-actions-powering-faster-more-efficient-build-systems/#get-started-using-arm-hosted-runners-today
Signed-off-by: Dennis Ameling <dennis@dennisameling.com>